+++ fpy2/runtime/profile/__init__.py
@@ -1,0 +1,2 @@
+from .expr_profiler import ExprProfiler, ErrorProfile
+from .function_profiler import FunctionProfiler
