+++ fpy2/runtime/real/__init__.py
@@ -1,4 +1,2 @@
 from .interpreter import RealInterpreter
 from .interval import Interval, BoolInterval, RealInterval
-from .function_profiler import FunctionProfiler
-from .expr_profiler import ExpressionProfiler
