+++ infra/real/expr_profiler.py
@@ -2,7 +2,6 @@ from typing import Optional
 
 from fpy2 import *
 from fpy2.runtime.sampling import sample_function
-from fpy2.runtime.real import ExpressionProfiler
 
 from .common import disabled_tests, select_interpreter
 from .config import Config
@@ -10,7 +9,7 @@ from .load import load_funs
 
 def _run_one(
     fun: Function,
-    profiler: ExpressionProfiler,
+    profiler: ExprProfiler,
     num_samples: int,
     *,
     seed: Optional[int] = None,
@@ -26,7 +25,7 @@ def _run_one(
 
 def run_expr_profiler(config: Config):
     reference = select_interpreter(config.ref_mode)
-    profiler = ExpressionProfiler(reference=reference, logging=True)
+    profiler = ExprProfiler(reference=reference, logging=True)
     funs = load_funs(config.input_paths)
 
     disabled = disabled_tests()
